android: add legacy system accent variant to fix LineageOS power menu reboot
For unknown reasons, when having enabled the sound panel in the power
menu on LineageOS (Substratum legacy/RRO, no issues e.g. on AICP
Substratum/OMS-supported), the overlaid drawables aren't found, leading
to a frameworks crash/soft reboot.

As a workaround, add a system accent variant that uses the same accent
as settings, but does not include the problematic drawables.

Change-Id: Ibd602c3211118af7b0b048fa76c2aefb773e005a